REF: Add `Modifier` to `ContentStatusText` param, move to `ui` package
