+++ app/src/main/java/cz/quanti/razym/rocketapp/system/RocketDetailScreen.kt
@@ -57,6 +57,7 @@ import cz.quanti.razym.rocketapp.R
 import cz.quanti.razym.rocketapp.presentation.RocketDetail
 import cz.quanti.razym.rocketapp.presentation.RocketDetailViewModel
 import cz.quanti.razym.rocketapp.presentation.Stage
+import cz.quanti.razym.rocketapp.ui.ContentStatusText
 import cz.quanti.razym.rocketapp.ui.theme.RocketappTheme
 import org.koin.androidx.compose.getViewModel
